chore: move models to model package and add mappers
