+++ android/native/emarsys/src/main/java/com/emarsys/maui/DotnetEmarsysGeofence.java
@@ -2,6 +2,9 @@ package com.emarsys.maui;
 
 import androidx.annotation.NonNull;
 import com.emarsys.Emarsys;
+import com.emarsys.maui.mapper.GeofenceMapper;
+import com.emarsys.maui.model.EMSGeofence;
+import com.emarsys.maui.model.EMSGeofenceTrigger;
 import com.emarsys.mobileengage.api.geofence.Geofence;
 import com.emarsys.mobileengage.api.geofence.Trigger;
 import java.util.ArrayList;
@@ -31,21 +34,7 @@ public class DotnetEmarsysGeofence {
 
     public static @NonNull List<EMSGeofence> getRegisteredGeofences() {
         List<Geofence> geofences = Emarsys.getGeofence().getRegisteredGeofences();
-        return mapGeofences(geofences);
-    }
-
-    static @NonNull List<EMSGeofence> mapGeofences(@NonNull List<Geofence> geofences) {
-        List<EMSGeofence> _geofences = new ArrayList<>();
-        for (int i = 0; i < geofences.size(); i++) {
-            Geofence g = geofences.get(i);
-            List<EMSGeofenceTrigger> triggers = new ArrayList<>();
-            for (int j = 0; j < g.getTriggers().size(); j++) {
-                Trigger t = g.getTriggers().get(j);
-                triggers.add(new EMSGeofenceTrigger(t.getId(), t.getType().toString(), t.getLoiteringDelay(), t.getAction()));
-            }
-            _geofences.add(new EMSGeofence(g.getId(), g.getLat(), g.getLon(), g.getRadius(), g.getWaitInterval(), triggers));
-        }
-        return _geofences;
+        return GeofenceMapper.mapGeofences(geofences);
     }
 
 }
