+++ android/native/emarsys/src/main/java/com/emarsys/maui/model/EMSConfig.java
@@ -1,4 +1,4 @@
-package com.emarsys.maui;
+package com.emarsys.maui.model;
 
 import android.app.Application;
 import androidx.annotation.NonNull;
